requirements.txt
Requirements:

numpy
panda
pyplot
matplotlib
